New system to move around the field
Use test mode to move around the field. You can move/turn the robot with the mouse.
Also, move functionality of MoveAroundField command into this routine.
In Robot modeInit methods, set robot position at the start.
In Robot.disabledPeriodic(), only move robot if the auto changes, so that it does
   not immediately override the testMode values.